consolidated modal handling, added prop comments, broke out large templates and common form elements
