+++ src/terra-dev-site/IssueForm/FeatureTemplate.jsx
@@ -1,0 +1,12 @@
+const FeatureTemplate = props => `# Feature Request
+
+## Description
+${props.description}
+
+${props.context ? `## Additional Context / Screenshots
+${props.context}` : ''}
+
+${props.mentions ? `## Mentions
+${props.mentions}` : ''}`;
+
+export default FeatureTemplate;
