+++ src/terra-dev-site/IssueForm/Helper.jsx
@@ -8,17 +8,45 @@ const propTypes = {
    */
   packageList: PropTypes.arrayOf(PropTypes.string),
   /**
-   * Functions that set state with the provided field values.
+   * Function that sets context state with the additional context field value.
    */
   setContext: PropTypes.func,
+  /**
+   * Function that sets description state with the description field value.
+   */
   setDescription: PropTypes.func,
+  /**
+   * Function that sets environment state with the environment field value.
+   * Note: This includes the pre-fill template.
+   */
   setEnvironment: PropTypes.func,
+  /**
+   * Function that sets expected state with the expected result field value.
+   */
   setExpected: PropTypes.func,
+  /**
+   * Function that sets issueType state with the issue type select value.
+   */
   setIssue: PropTypes.func,
+  /**
+   * Function that sets mentions state with the mentions field value.
+   */
   setMentions: PropTypes.func,
+  /**
+   * Function that sets selectedPackage state with the package select value.
+   */
   setPackage: PropTypes.func,
+  /**
+   * Function that sets solution state with the potential solution field value.
+   */
   setSolution: PropTypes.func,
+  /**
+   * Function that sets steps state with the steps to reproduce field value.
+   */
   setSteps: PropTypes.func,
+  /**
+   * Function that sets title state with the title field value.
+   */
   setTitle: PropTypes.func,
   /**
    * Sets the value of the issue type select box to the selected issue type.
@@ -27,50 +55,16 @@ const propTypes = {
 };
 
 // Templates used for preview, errors, and forms.
-const bugTemplate = props => `# Bug Report
-
-## Description
-${props.description}
-
-## Steps to Reproduce
-${props.steps}
-
-## Expected Behavior
-${props.expected}
-
-## Environment
-${props.environment}
-
-${props.context ? `## Additional Context / Screenshots
-${props.context}` : ''}
-
-${props.solution ? `## Possible Solution
-${props.solution}` : ''}
-
-${props.mentions ? `## Mentions
-${props.mentions}` : ''}`;
-
 const disclaimerTemplate = `When logging issues to Terra UI, it is helpful to have steps to reproduce the issue using a reduced test case, or examples from <a href="https://engineering.cerner.com/terra-ui" target="_blank">https://engineering.cerner.com/terra-ui</a>.
     If there are videos or screenshots that are available, those are helpful as well.
 
 **Note to Cerner associates:** If you have proprietary information, don't post it directly.  Instead post a link to a site that can be protected by the Cerner IDP (e.g. sharepoint).`;
 
-const environmentTemplate = `* Component Version:
-* Browser Name and Version:
-* Node/npm Version [e.g. Node 8/npm 5]:
-* Webpack Version:
+const environmentTemplate = `* Component Version: 
+* Browser Name and Version: 
+* Node/npm Version [e.g. Node 8/npm 5]: 
+* Webpack Version: 
 * Operating System and version (desktop or mobile): `;
-
-const featureTemplate = props => `# Feature Request
-
-## Description
-${props.description}
-
-${props.context ? `## Additional Context / Screenshots
-${props.context}` : ''}
-
-${props.mentions ? `## Mentions
-${props.mentions}` : ''}`;
 
 const titleTemplate = (title, repo, selectedPackage) => `# Repo
 ${repo}
@@ -105,10 +99,8 @@ const validateForm = (value) => {
 };
 
 export {
-  bugTemplate,
   environmentTemplate,
   disclaimerTemplate,
-  featureTemplate,
   getPackages,
   getRepo,
   propTypes,
